191007 | 오늘은~~

+++ 03_Django/04_django_crud_review/articles/views.py
@@ -1,11 +1,13 @@
 from django.shortcuts import render, redirect
 from django.core.exceptions import ValidationError
 from .models import Article
+from IPython import embed
 
 
 def index(request):
     # articles = Article.objects.all()[::-1]
     articles = Article.objects.order_by('-pk')
+    embed()
     # print(articles)
     # print(type(articles))
     context = {
@@ -13,20 +15,31 @@ def index(request):
     }
     return render(request, 'articles/index.html', context)
 
-def new(request):
-    return render(request, 'articles/new.html')
-
 def create(request):
-    try:
+    # POST 요청일 때
+    if request.method == 'POST':
         title = request.POST.get('title')
         content = request.POST.get('content')
         article = Article(title=title, content=content)
-        article.full_clean()
-    except ValidationError:
-        raise ValidationError('Your Error Message')
-    else:
         article.save()
         return redirect('articles:detail', article.pk)
+    # GET 요청일 때
+    else:
+        return render(request, 'articles/create.html')
+
+    # try:
+    #     title = request.POST.get('title')
+    #     content = request.POST.get('content')
+    #     article = Article(title=title, content=content)
+    #     article.full_clean()
+    # except ValidationError:
+    #     raise ValidationError('Your Error Message')
+    # else:
+    #     article.save()
+    #     return redirect('articles:detail', article.pk)
+
+
+    
     '''
     title = request.POST.get('title')
     content = request.POST.get('content')
@@ -50,6 +63,7 @@ def create(request):
 
 def detail(request, pk):
     article = Article.objects.get(pk=pk)
+    embed()
     context = {
         'article':article
     }
@@ -57,20 +71,29 @@ def detail(request, pk):
 
 def delete(request, pk):
     article = Article.objects.get(pk=pk)
-    article.delete()
-    return redirect('articles:index')
+    if request.method == 'POST':
+        article.delete()
+        return redirect('articles:index')
+    return redirect('articles:detail', article.pk)
 
-def edit(request, pk): 
-    article = Article.objects.get(pk=pk)
-    context = {
-        'article':article
-    }
-    return render(request, 'articles/edit.html', context)
+# def edit(request, pk): 
+#     article = Article.objects.get(pk=pk)
+#     context = {
+#         'article':article
+#     }
+#     return render(request, 'articles/edit.html', context)
 
 def update(request, pk):
+    # embed()
     article = Article.objects.get(pk=pk)
-    article.title = request.POST.get('title')
-    article.content = request.POST.get('content')
-    article.save()
-
-    return redirect('articles:detail', article.pk)
+    if request.method == 'POST':
+        article.title = request.POST.get('title')
+        article.content = request.POST.get('content')
+        article.save()
+        return redirect('articles:detail', article.pk)
+        
+    else:       
+        context = {
+            'article':article
+        }
+        return render(request, 'articles/update.html', context)
